#77 devide some helper methods
